Removed dependence on fuse and ntfs3g
Removed dependence on fuse and ntfs3g as this tool does not require to mount WIM images. NTFS3g is probably also not needed as images are not directly applied to an NTFS volume. But rather being extracted into an already existing one.
However as development progresses this might change.